Rewrite QADataset to take just a simple dataframe
Before, initializing a `QADataset` required passing a catalog and a group of functors, and the QADataset object would manage the computation of the functor columns on that dataset.  This commit changes QADataset to simply receive an already-calculated dataframe, taking away all calculation functionality, and just managing the creation and wrapping of the holoviews `Dataset` object; e.g., by hiding away things like defining the kdims and vdims, getting rid of null values, etc.

This commit also keeps the old version of dataset.py as dataset_old.py.
This is a temporary convenience to keep all the old stuff visible in the
current repo until all the old functionality is replicated, at which
point it will be removed.
